+++ webpack.config.js
@@ -2,20 +2,24 @@ var path = require('path');
 var webpack = require('webpack');
 
 module.exports = {
-    entry: './assets/main.js',
+    entry: {
+        bundle: './assets/main.js',
+        initial: './assets/initial.scss'
+    },
     output: {
-        filename: 'bundle.js',
+        filename: '[name].js',
         path: path.resolve(__dirname, 'public')
     },
     module: {
         rules: [{
             test: /\.css$/,
-            use: [ 'style-loader', 'css-loader' ]
+            use: ['style-loader', 'css-loader', 'postcss-loader']
         },{
             test: /\.scss$/,
             use: [
                 'style-loader',
                 'css-loader',
+                'postcss-loader',
                 'sass-loader',
                 'image-webpack-loader'
             ]