+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Add build directory to .gitignore
 
+### Changed
+- Update README.md information
+
 ## [v0.0.1] - 28.09.17
 ### Added
 - Add build step with pkg