+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.0.3] - 09.10.17
 ### Added
 - Add cleanup call on every request and by configured interval
 - Add template inheritance for theming
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fix condition race for activeuntil on show message page
-- Fix visible scrollbars if hardware mouse detected
+- Fix visible scrollbars if hardware mouse detected #1
 
 ## [v0.0.2] - 30.09.17
 ### Added